Handle special entries and unusual position values #20
Fixes two bugs in imenu-list--current-entry:

- special entries: the code didn't extract the position parameter
correctly from special entries ("special" as defined in
imenu--index-alist's docstring)

- unusual position values: position parameter doesn't strictly have to
be a number or a marker, though it almost always is.
imenu-list-position-translator returns a translation function from
position paramter to number/marker.
imenu-list-custom-position-translator allows the user to add translation
rules that are missing from imenu-list.